Move `StackContents` to its own file.
